Adding local user managment
* keys-sync user becomes a local and not an ldap user
* local users are allowed to access site without 403
* LDAP can be dis/enabled in the configuration
* Added section to delete local user
* Added section to add local users
* Added graceful handling of deleted users in logs, nodes, ...

There are tables containing NOT NULL fields for actors
like the log table, which contains who did something.
When we remove a user, we don't want to remove that log
as the entity it is for is still alive. We instead want
to remove the actor (or set it to null) so that the
datbase stays consitent. The migrations in this patch
recreate every table which has an actor to allow NULL
fields. It also corrects a few constraints to set those
values to NULL on delete.